requirements.txt
Adafruit-Blinka==8.64.0
adafruit-circuitpython-busdevice==5.2.13
adafruit-circuitpython-connectionmanager==3.1.5
adafruit-circuitpython-neopixel==6.3.9
adafruit-circuitpython-pixelbuf==2.0.9
adafruit-circuitpython-requests==4.1.13
adafruit-circuitpython-typing==1.12.1
Adafruit-PlatformDetect==3.82.0
Adafruit-PureIO==1.1.11
bidict==0.23.1
binho-host-adapter==0.1.6
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorzero==2.0
contourpy==1.3.3
cycler==0.12.1
filelock==3.19.1
filterpy==1.4.5
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
fonttools==4.59.1
fsspec==2025.7.0
gpiozero==1.6.2
h11==0.16.0
idna==3.10
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.7.2
mpmath==1.3.0
networkx==3.5
numpy==1.25.2
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
packaging==25.0
pandas==2.0.3
Pillow==10.0.0
psutil==7.0.0
py-cpuinfo==9.0.0
pyftdi==0.57.1
pyparsing==3.0.9
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-engineio==4.12.2
python-multipart==0.0.6
python-socketio==5.13.0
pytz==2025.2
pyusb==1.3.1
PyWavelets==1.9.0
PyYAML==6.0.1
requests==2.31.0
RPi.GPIO==0.7.1
rpi_ws281x==4.3.4
scikit-image==0.21.0
scipy==1.16.1
seaborn==0.13.2
simple-websocket==1.1.0
six==1.17.0
smbus2==0.4.2
sympy==1.14.0
sysv_ipc==1.1.0
thop==0.1.1.post2209072238
tifffile==2025.6.11
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.0.196
urllib3==2.5.0
Werkzeug==3.1.3
wsproto==1.2.0
